Start of rewrite.
Now, we target UEFI directly, and have a QEMU script to run, downloading
the firmware off of TFTP.

This will require revisiting later when attempting to run on real
hardware, as we have to write a bootloader to DL the OS image from the
TFTP server as UEFI does. but that is later.